@@ -30,14 +30,13 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  google_maps_flutter: ^2.2.0
-  geolocator: ^10.0.1
-
+  webview_flutter: ^4.0.0 # Pastikan ini adalah versi terbaru
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
   url_launcher: ^6.3.1
+  geolocator: ^14.0.0
 
 dev_dependencies:
   flutter_test:
