@@ -13,21 +13,23 @@ class _SitusRekomendasiPageState extends State<SitusRekomendasiPage> {
     {
       'title': 'Google',
       'url': 'https://www.google.com',
-      'image': 'https://www.google.com/favicon.ico'
+      'image': 'https://www.google.com/favicon.ico',
     },
     {
       'title': 'Flutter',
       'url': 'https://flutter.dev',
-      'image': 'https://flutter.dev/assets/images/shared/brand/flutter/logo/flutter-lockup.png'
+      'image':
+          'https://flutter.dev/assets/images/shared/brand/flutter/logo/flutter-lockup.png',
     },
     {
       'title': 'GitHub',
       'url': 'https://github.com',
-      'image': 'https://github.githubassets.com/images/modules/logos_page/GitHub-Mark.png'
+      'image':
+          'https://github.githubassets.com/images/modules/logos_page/GitHub-Mark.png',
     },
   ];
 
-  final List<Map<String, String>> _favoriteSitus = [];
+  final Set<Map<String, String>> _favoriteSitus = {};
 
   void _toggleFavorite(Map<String, String> situs) {
     setState(() {
@@ -40,12 +42,13 @@ class _SitusRekomendasiPageState extends State<SitusRekomendasiPage> {
   }
 
   void _launchURL(String url) async {
-    if (await canLaunch(url)) {
-      await launch(url);
+    final Uri uri = Uri.parse(url);
+    if (await canLaunchUrl(uri)) {
+      await launchUrl(uri, mode: LaunchMode.externalApplication);
     } else {
-      ScaffoldMessenger.of(context).showSnackBar(
-        SnackBar(content: Text('Tidak dapat membuka link')),
-      );
+      ScaffoldMessenger.of(
+        context,
+      ).showSnackBar(const SnackBar(content: Text('Tidak dapat membuka link')));
     }
   }
 
@@ -55,18 +58,27 @@ class _SitusRekomendasiPageState extends State<SitusRekomendasiPage> {
       length: 2,
       child: Scaffold(
         appBar: AppBar(
-          title: Text('Situs Rekomendasi'),
-          bottom: TabBar(
-            tabs: [
-              Tab(text: 'Semua Situs'),
-              Tab(text: 'Favorite ❤️'),
-            ],
+          title: const Text('Situs Rekomendasi'),
+          bottom: const TabBar(
+            labelColor: Color.fromARGB(
+              255,
+              254,
+              254,
+              254,
+            ), // Warna teks tab yang aktif
+            unselectedLabelColor: Color.fromARGB(
+              255,
+              188,
+              182,
+              182,
+            ), // Warna teks tab yang tidak aktif
+            tabs: [Tab(text: 'Semua Situs'), Tab(text: 'Favorite ❤️')],
           ),
         ),
         body: TabBarView(
           children: [
             _buildSitusList(_situs),
-            _buildSitusList(_favoriteSitus),
+            _buildSitusList(_favoriteSitus.toList()),
           ],
         ),
       ),
@@ -80,12 +92,21 @@ class _SitusRekomendasiPageState extends State<SitusRekomendasiPage> {
         final situs = situsList[index];
         return Card(
           child: ListTile(
-            leading: Image.network(situs['image']!, width: 40, height: 40, errorBuilder: (context, error, stackTrace) => Icon(Icons.image)),
+            leading: Image.network(
+              situs['image']!,
+              width: 40,
+              height: 40,
+              errorBuilder:
+                  (context, error, stackTrace) =>
+                      const Icon(Icons.broken_image),
+            ),
             title: Text(situs['title']!),
             subtitle: Text(situs['url']!),
             trailing: IconButton(
               icon: Icon(
-                _favoriteSitus.contains(situs) ? Icons.favorite : Icons.favorite_border,
+                _favoriteSitus.contains(situs)
+                    ? Icons.favorite
+                    : Icons.favorite_border,
                 color: _favoriteSitus.contains(situs) ? Colors.red : null,
               ),
               onPressed: () => _toggleFavorite(situs),
