@@ -1,9 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:tugas3/pages/our_team.dart';
-// import 'package:tugas3/pages/members_page.dart';
 import 'package:tugas3/pages/help_page.dart';
-// import 'package:tugas3/main.dart';
-
 import 'package:tugas3/pages/stopwatch_page.dart';
 import 'package:tugas3/pages/jenis_bilangan_page.dart';
 import 'package:tugas3/pages/tracking_page.dart';
@@ -20,11 +17,10 @@ class HomePage extends StatefulWidget {
 class _HomePageState extends State<HomePage> {
   int _selectedIndex = 0;
 
-  // Daftar halaman dalam Bottom Navigation Bar
   final List<Widget> _pages = [
-    MainMenuPage(),   // Halaman Utama
-    OurTeamPage(),     // Daftar Anggota
-    HelpPage(),       // Bantuan & Logout
+    const MainMenuPage(), // Halaman Utama
+    const OurTeamPage(), // Daftar Anggota
+    const HelpPage(), // Bantuan & Logout
   ];
 
   void _onItemTapped(int index) {
@@ -35,91 +31,120 @@ class _HomePageState extends State<HomePage> {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: IndexedStack(
-        index: _selectedIndex,
-        children: _pages,
+    return MaterialApp(
+      debugShowCheckedModeBanner: false,
+      title: 'Tugas 3',
+      theme: ThemeData(
+        primarySwatch: Colors.purple,
+        scaffoldBackgroundColor: Colors.purple[50], // Ubah ke ungu biasa
+        appBarTheme: const AppBarTheme(
+          backgroundColor: Colors.purple, // Ubah ke ungu biasa
+          foregroundColor: Colors.white,
+          centerTitle: true,
+        ),
       ),
-      bottomNavigationBar: BottomNavigationBar(
-        currentIndex: _selectedIndex,
-        onTap: _onItemTapped,
-        items: [
-          BottomNavigationBarItem(icon: Icon(Icons.home), label: 'Home'),
-          BottomNavigationBarItem(icon: Icon(Icons.group), label: 'Anggota'),
-          BottomNavigationBarItem(icon: Icon(Icons.help), label: 'Bantuan'),
-        ],
+      home: Scaffold(
+        body: IndexedStack(index: _selectedIndex, children: _pages),
+        bottomNavigationBar: BottomNavigationBar(
+          currentIndex: _selectedIndex,
+          onTap: _onItemTapped,
+          items: const [
+            BottomNavigationBarItem(icon: Icon(Icons.home), label: 'Home'),
+            BottomNavigationBarItem(icon: Icon(Icons.group), label: 'Anggota'),
+            BottomNavigationBarItem(icon: Icon(Icons.help), label: 'Bantuan'),
+          ],
+        ),
       ),
     );
   }
 }
 
-// Halaman Utama (Berisi daftar fitur)
 class MainMenuPage extends StatelessWidget {
   const MainMenuPage({super.key});
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: Text('Halaman Utama')),
+      appBar: AppBar(title: const Text('Halaman Utama')),
       body: ListView(
-        padding: EdgeInsets.all(16),
+        padding: const EdgeInsets.all(16.0),
         children: [
-          ListTile(
-            leading: Icon(Icons.timer),
-            title: Text('Stopwatch'),
-            onTap: () {
-              Navigator.push(
-                context,
-                MaterialPageRoute(builder: (context) => StopwatchPage()),
-              );
-            },
+          _buildCard(
+            context,
+            title: 'Stopwatch',
+            icon: Icons.timer,
+            page: StopwatchPage(),
           ),
-          // Tambahkan fitur lainnya di sini
-          ListTile(
-            leading: Icon(Icons.format_list_numbered),
-            title: Text('Cek Jenis Bilangan'),
-            onTap: () {
-              Navigator.push(
-                context,
-                MaterialPageRoute(builder: (context) => JenisBilanganPage()),
-              );
-            },
+          _buildCard(
+            context,
+            title: 'Jenis Bilangan',
+            icon: Icons.calculate,
+            page: JenisBilanganPage(),
           ),
-          ListTile(
-            leading: Icon(Icons.map),
-            title: Text('Tracking LBS'),
-            onTap: () {
-              Navigator.push(
-                context,
-                MaterialPageRoute(builder: (context) => TrackingPage()),
-              );
-            },
+          _buildCard(
+            context,
+            title: 'Tracking LBS',
+            icon: Icons.location_on,
+            page: TrackingPage(),
           ),
-          ListTile(
-              leading: Icon(Icons.access_time),
-              title: Text('Konversi Waktu'),
-              onTap: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(builder: (context) => KonversiWaktuPage()),
-                );
-              },
-            ),
-          ListTile(
-            leading: Icon(Icons.web),
-            title: Text('Situs Rekomendasi'),
-            onTap: () {
-              Navigator.push(
-                context,
-                MaterialPageRoute(builder: (context) => SitusRekomendasiPage()),
-              );
-            },
+          _buildCard(
+            context,
+            title: 'Konversi Waktu',
+            icon: Icons.access_time,
+            page: KonversiWaktuPage(),
           ),
-
-
+          _buildCard(
+            context,
+            title: 'Favorit',
+            icon: Icons.favorite,
+            page: SitusRekomendasiPage(),
+          ),
         ],
       ),
     );
   }
-}
 
+  Widget _buildCard(
+    BuildContext context, {
+    required String title,
+    required IconData icon,
+    required Widget page,
+  }) {
+    return InkWell(
+      onTap:
+          () => Navigator.push(
+            context,
+            MaterialPageRoute(builder: (context) => page),
+          ),
+      splashColor: Colors.purple.withOpacity(0.3),
+      borderRadius: BorderRadius.circular(12),
+      child: Card(
+        margin: const EdgeInsets.symmetric(vertical: 10),
+        elevation: 5,
+        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
+        child: Padding(
+          padding: const EdgeInsets.all(16.0),
+          child: Row(
+            children: [
+              CircleAvatar(
+                backgroundColor: Colors.deepPurple.shade100,
+                child: Icon(icon, color: Colors.deepPurple),
+              ),
+              const SizedBox(width: 20),
+              Expanded(
+                child: Text(
+                  title,
+                  style: const TextStyle(
+                    fontSize: 18,
+                    fontWeight: FontWeight.w600,
+                  ),
+                ),
+              ),
+              const Icon(Icons.arrow_forward_ios, color: Colors.deepPurple),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
