@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:tugas3/main.dart';
+import 'package:tugas3/pages/home_page.dart';
 
 class OurTeamPage extends StatelessWidget {
   const OurTeamPage({super.key});
@@ -13,32 +15,49 @@ class OurTeamPage extends StatelessWidget {
         leading: IconButton(
           icon: Icon(Icons.arrow_back, color: Colors.black),
           onPressed: () {
-            Navigator.pop(context);
+            Navigator.pushReplacement(
+              context,
+              MaterialPageRoute(
+                builder: (context) => HomePage(),
+              ), // Kembali ke MainPage
+            );
           },
         ),
       ),
-      body: Column(
-        children: [
-          Padding(
-            padding: const EdgeInsets.all(16.0),
-            child: Text(
-              "OUR TEAM",
-              style: TextStyle(fontSize: 24, fontWeight: FontWeight.bold),
+      body: SingleChildScrollView(
+        // Tambahkan SingleChildScrollView
+        child: Column(
+          children: [
+            Padding(
+              padding: const EdgeInsets.all(16.0),
+              child: Text(
+                "OUR TEAM",
+                style: TextStyle(fontSize: 24, fontWeight: FontWeight.bold),
+              ),
             ),
-          ),
-          Expanded(
-            child: Column(
+            Column(
+              // Hapus Expanded di sini
               children: [
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
+                Row(mainAxisAlignment: MainAxisAlignment.center),
+                _buildTeamMember(
+                  "assets/profile3.jpg",
+                  "RAFIQ IKHWAN NUGRAHA",
+                  "123220071",
                 ),
-                _buildTeamMember("assets/profile3.jpg", "RAFIQ IKHWAN NUGRAHA", "123220071"),
-                _buildTeamMember("assets/profile.jpg", "SYIFA NUR RAMADHANI", "123220194"),
-                _buildTeamMember("assets/profile2.jpg", "MUH. NAUFAL FAUZI ALI", "123220207"),
+                _buildTeamMember(
+                  "assets/profile.jpg",
+                  "SYIFA NUR RAMADHANI",
+                  "123220194",
+                ),
+                _buildTeamMember(
+                  "assets/profile2.jpg",
+                  "MUH. NAUFAL FAUZI ALI",
+                  "123220207",
+                ),
               ],
             ),
-          ),
-        ],
+          ],
+        ),
       ),
     );
   }
@@ -48,10 +67,7 @@ class OurTeamPage extends StatelessWidget {
       padding: const EdgeInsets.all(8.0),
       child: Column(
         children: [
-          CircleAvatar(
-            radius: 50,
-            backgroundImage: AssetImage(imagePath),
-          ),
+          CircleAvatar(radius: 50, backgroundImage: AssetImage(imagePath)),
           SizedBox(height: 10),
           Container(
             padding: EdgeInsets.symmetric(vertical: 8, horizontal: 16),
@@ -61,10 +77,7 @@ class OurTeamPage extends StatelessWidget {
             ),
             child: Column(
               children: [
-                Text(
-                  name,
-                  style: TextStyle(fontWeight: FontWeight.bold),
-                ),
+                Text(name, style: TextStyle(fontWeight: FontWeight.bold)),
                 Text(id),
               ],
             ),
